Moved NewV7FromReader to uuid_test since it shouldn't be used outside of testing
